requirements.txt
fastapi
uvicorn
pymongo
passlib[bcrypt]
python-jwt   # or “pyjwt”
